requirements.txt
account==0.1.0
accounts==0.2.4
activate==0.2.1
alembic==1.11.1
amqp==2.6.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
billiard==3.6.4.0
cachetools==5.3.1
celery==4.4.7
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
chardet==5.1.0
charset-normalizer==3.1.0
click==7.1.2
cryptography==41.0.1
cssselect==1.2.0
cssutils==2.7.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
djangorestframework==3.14.0
dnspython==2.3.0
ecdsa==0.18.0
elastic-transport==8.4.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
email-validator==1.3.1
emails==0.5.15
exceptiongroup==1.1.1
fastapi==0.70.1
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
gunicorn==20.1.0
h11==0.9.0
httptools==0.1.2
idna==3.4
inflect==5.6.2
Jinja2==2.11.3
kombu==4.6.11
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
more-itertools==9.1.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
passlib==1.7.4
Pillow==9.5.0
pluggy==0.13.1
premailer==3.10.0
protobuf==4.23.4
psycopg2-binary==2.9.6
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.10
PyJWT==2.7.0
pymongo==4.4.1
pytest==5.4.3
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2023.3
raven==6.10.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
sqlparse==0.4.4
starlette==0.16.0
tenacity==6.3.1
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.11.8
uvloop==0.17.0
vine==1.3.0
wcwidth==0.2.6
websockets==8.1
